Moving the harvest unit test to the correct location.
